fix: improve API literacy
