+++ novaan-server/src/Content/ContentController.cs
@@ -61,7 +61,7 @@ namespace NovaanServer.src.Content
             return Ok();
         }
 
-        
+
     }
 }
